refactor: decompose module target preparation
